refac bdd tests
