+++ src/test/java/org/delia/app/DeliaDaoTests.java
@@ -4,7 +4,7 @@ package org.delia.app;
 import static org.junit.Assert.*;
 
 import org.delia.api.Delia;
-import org.delia.bddnew.NewBDDBase;
+import org.delia.bdd.NewBDDBase;
 import org.delia.builder.ConnectionInfo;
 import org.delia.builder.DeliaBuilder;
 import org.delia.builder.ConnectionBuilder;
