+++ src/test/java/org/delia/bdd/NewBDDH2Tests.java
@@ -1,13 +1,14 @@
-package org.delia.bddnew;
+package org.delia.bdd;
 
-import org.delia.bddnew.core.BDDTester2;
+import org.delia.bdd.core.BDDTester2;
+import org.delia.bdd.core.MyFakeSQLDBInterface;
 import org.delia.db.DBInterface;
-import org.delia.db.memdb.MemDBInterface;
+import org.delia.db.DBType;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
 
-public class AllBDDTests extends NewBDDBase {
+public class NewBDDH2Tests extends NewBDDBase {
 	
 	//TODO: R100
 	//TODO: R200
@@ -33,9 +34,9 @@ public class AllBDDTests extends NewBDDBase {
 	public void testR400() {
 		runR400File("t0-struct-key.txt", 9);
 		runR400File("t0-field-unique.txt", 6);
-		runR400File("t0-field-optional.txt", 4);
 		runR400File("t0-struct.txt", 4);
 		runR400File("t0-struct-inheritance.txt", 6);
+		runR400File("t0-field-optional.txt", 4);
 		runR400File("t0-field-primarykey.txt", 6);
 		runR400File("t0-field-serial.txt", 7);
 	}
@@ -75,7 +76,6 @@ public class AllBDDTests extends NewBDDBase {
 	@Test
 	public void testR700() {
 		runR700File("t0-insert.txt", 6);
-		runR700File("t0-insert-serial.txt", 1);
 	}
 	
 	@Test
@@ -98,7 +98,7 @@ public class AllBDDTests extends NewBDDBase {
 	public void testR950() {
 		runR950File("t0-crud-assoc-insert.txt", 0);
 	}
-
+	
 	@Test
 	public void testR1000() {
 		runR1000File("t0-upsert.txt", 5);
@@ -181,12 +181,13 @@ public class AllBDDTests extends NewBDDBase {
 		runR1550File("t0-queryfn-oneone-parent.txt", 6);
 		runR1550File("t0-queryfn-oneone-parent2.txt", 6);
 		runR1550File("t0-queryfn-oneone-child.txt", 6);
-		runR1550File("t0-queryfn-oneone-childa.txt", 7);
+		//runR1550File("t0-queryfn-oneone-childa.txt", 6);
 		runR1550File("t0-queryfn-onemany-parent.txt", 6);
 		runR1550File("t0-queryfn-onemany-child.txt", 6);
 		runR1550File("t0-queryfn-manymany-left.txt", 6);
 		runR1550File("t0-queryfn-manymany-right.txt", 6);
 	}	
+	
 	@Test
 	public void testR1600() {
 		runR1600File("t0-fetch.txt", 3);
@@ -200,7 +201,7 @@ public class AllBDDTests extends NewBDDBase {
 	public void testR1700() {
 		runR1700File("t0-let-field-single.txt", 3);
 		runR1700File("t0-let-field-func.txt", 3);
-		runR1700File("t0-let-field-multiple.txt", 5); 
+		runR1700File("t0-let-field-multiple.txt", 5);
 		runR1700File("t0-let-field-relation.txt", 2);
 	}
 	
@@ -221,34 +222,31 @@ public class AllBDDTests extends NewBDDBase {
 	
 	@Test
 	public void testR2100() {
-		enableAllFileCheck = false;
 		runR2100File("t0-migration.txt", 2);
 		runR2100File("t0-migration2.txt", 2);
 		runR2100File("t0-migration3.txt", 2);
-		//TODO: this don't work on mem db. fix
-//		runR2100File("t0-migration3a.txt", 2);
-//		runR2100File("t0-migration3b.txt", 2);
-//		runR2100File("t0-migration3c.txt", 2);
-//		runR2100File("t0-migration4.txt", 2);
-//		runR2100File("t0-migration10.txt", 2);
-//		runR2100File("t0-migration10a.txt", 2);
-//		runR2100File("t0-migration10b.txt", 3);
-//		runR2100File("t0-migration11.txt", 0);
-//		runR2100File("t0-migration5.txt", 0);
-//		runR2100File("t0-migration7.txt", 2);
-//		runR2100File("t0-migration12.txt", 2);
-//		runR2100File("t0-migration13.txt", 2);
-//		runR2100File("t0-migration13a.txt", 2);
-//		runR2100File("t0-migration14.txt", 2);
-//		runR2100File("t0-migration14a.txt", 2);
-//		runR2100File("t0-migration15.txt", 2);
-//		runR2100File("t0-migration15a.txt", 2);
-//		runR2100File("t0-migration16.txt", 2);
-//		runR2100File("t0-migration16a.txt", 2);
+		runR2100File("t0-migration3a.txt", 2);
+		runR2100File("t0-migration3b.txt", 2);
+		runR2100File("t0-migration3c.txt", 2);
+		runR2100File("t0-migration4.txt", 2);
+		runR2100File("t0-migration10.txt", 2);
+		runR2100File("t0-migration10a.txt", 2);
+		runR2100File("t0-migration10b.txt", 3);
+		runR2100File("t0-migration11.txt", 0);
+		runR2100File("t0-migration5.txt", 2);
+		runR2100File("t0-migration7.txt", 2);
+		runR2100File("t0-migration12.txt", 2);
+		runR2100File("t0-migration13.txt", 2);
+		runR2100File("t0-migration13a.txt", 2);
+		runR2100File("t0-migration14.txt", 2);
+		runR2100File("t0-migration14a.txt", 2);
+		runR2100File("t0-migration15.txt", 2);
+		runR2100File("t0-migration15a.txt", 2);
+		runR2100File("t0-migration16.txt", 2);
+		runR2100File("t0-migration16a.txt", 2);
 	}
 	@Test
 	public void testR2150() {
-		enableAllFileCheck = false;
 		runR2150File("t0-migrate-one-to-one1.txt", 3);
 		runR2150File("t0-migrate-one-to-one2.txt", 2);
 		runR2150File("t0-migrate-one-to-one2a.txt", 2);
@@ -270,12 +268,10 @@ public class AllBDDTests extends NewBDDBase {
 		runR2150File("t0-migrate-many-to-many2a.txt", 2);
 		runR2150File("t0-migrate-many-to-many3.txt", 2);
 		runR2150File("t0-migrate-many-to-many4.txt", 2);
-		//TODO: fix these
-		//runR2150File("t0-migrate-many-to-many6.txt", 1);
-		//runR2150File("t0-migrate-many-to-many6a.txt", 2);
-		//runR2150File("t0-migrate-many-to-many7.txt", 2);
+		runR2150File("t0-migrate-many-to-many6.txt", 2);
+		runR2150File("t0-migrate-many-to-many6a.txt", 2);
+		runR2150File("t0-migrate-many-to-many7.txt", 0);
 	}
-	
 	@Test
 	public void testR2200() {
 		runR2200File("t0-security-sql-injection.txt", 3);
@@ -284,37 +280,57 @@ public class AllBDDTests extends NewBDDBase {
 	public void testR2300() {
 		runR2300File("t0-multi-relation.txt", 0);
 	}
-	
-	
+
 	@Test
 	public void test8Debug() {
-//		testIndexToRun = 1;
-		enableAllFileCheck = false;
+//		testIndexToRun = 3;
 		BDDTester2.disableSQLLoggingDuringSchemaMigration = false;
-		enableMigration = true;
-
-//		runR1350File("t0-filter-op-relation.txt", 13);
-//		runR1000File("t0-upsert-mm-other.txt", 1);
-		runR1000File("t0-upsert-mm-all.txt", 1);
+		enableAllFileCheck = false;
+		enableSQLLogging = true;
+		cleanTables = true;
+		
+//		runR1550File("t0-queryfn-manymany-right.txt", 6);
+//		runR1000File("t0-upsert-mm-id.txt", 4);
+//		runR900File("t0-update-mm-id.txt", 4);
+		runR1000File("t0-upsert.txt", 5);
 	}
 	
 	//---
-	private boolean enableMigration;
-
+	private DBType dbType = DBType.H2;
+	private boolean cleanTables = true;
+	private boolean enableSQLLogging = true;
+	private boolean useFragmentParser = true;
+	
 	@Before
 	public void init() {
+		this.disableAllSlowTests = true;
+		disableAllSlowTestsIfNeeded();
 	}
 	@After
 	public void shutdown() {
 		chkAllFiles();
 	}
+
+	@Override
+	protected int runBDDFile(BDDGroup group, String filename, int numTests) {
+		MyFakeSQLDBInterface db = new MyFakeSQLDBInterface(dbType);
+		db.cleanTables = cleanTables;
+		db.useFragmentParser = useFragmentParser;
+		dbInterfaceToUse = db;
+//		DeliaClient.forcedDBInterface = db;
+		if (enableSQLLogging) {
+			dbInterfaceToUse.enableSQLLogging(true);
+		}
+		
+		return super.runBDDFile(group, filename, numTests);
+	}
+	
 	@Override
 	public DBInterface createForTest() {
-		MemDBInterface db = new MemDBInterface();
-		if (enableMigration) {
-			db.getCapabilities().setRequiresSchemaMigration(true);
-		}
+		MyFakeSQLDBInterface db = new MyFakeSQLDBInterface(dbType);
+		db.cleanTables = cleanTables;
+		dbInterfaceToUse = db;
 		return db;
 	}
-
+	
 }
