+++ src/test/java/org/delia/bdd/core/BDDTest.java
@@ -1,4 +1,4 @@
-package org.delia.bddnew.core;
+package org.delia.bdd.core;
 
 import java.util.ArrayList;
 import java.util.List;
