+++ src/test/java/org/delia/bdd/core/BDDTester2.java
@@ -1,4 +1,4 @@
-package org.delia.bddnew.core;
+package org.delia.bdd.core;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotNull;
@@ -8,15 +8,15 @@ import java.util.List;
 import org.apache.commons.lang3.StringUtils;
 import org.delia.api.DeliaSession;
 import org.delia.base.UnitTestLog;
-import org.delia.bddnew.core.checker.BDDResult;
-import org.delia.bddnew.core.checker.BoolChecker;
-import org.delia.bddnew.core.checker.DateChecker;
-import org.delia.bddnew.core.checker.IntChecker;
-import org.delia.bddnew.core.checker.LongChecker;
-import org.delia.bddnew.core.checker.StringChecker;
-import org.delia.bddnew.core.checker.StructChecker;
-import org.delia.bddnew.core.checker.ValueChecker;
-import org.delia.bddnew.core.checker.ValueCheckerBase;
+import org.delia.bdd.core.checker.BDDResult;
+import org.delia.bdd.core.checker.BoolChecker;
+import org.delia.bdd.core.checker.DateChecker;
+import org.delia.bdd.core.checker.IntChecker;
+import org.delia.bdd.core.checker.LongChecker;
+import org.delia.bdd.core.checker.StringChecker;
+import org.delia.bdd.core.checker.StructChecker;
+import org.delia.bdd.core.checker.ValueChecker;
+import org.delia.bdd.core.checker.ValueCheckerBase;
 import org.delia.client.ClientTests.DeliaClient;
 import org.delia.db.DBInterface;
 import org.delia.db.memdb.MemDBInterface;
