+++ src/test/java/org/delia/bdd/core/DBInterfaceCreator.java
@@ -1,4 +1,4 @@
-package org.delia.bddnew.core;
+package org.delia.bdd.core;
 
 import org.delia.db.DBInterface;
