+++ src/test/java/org/delia/bdd/core/checker/ValueChecker.java
@@ -1,7 +1,7 @@
-package org.delia.bddnew.core.checker;
+package org.delia.bdd.core.checker;
 
 import org.delia.api.DeliaSession;
-import org.delia.bddnew.core.ThenValue;
+import org.delia.bdd.core.ThenValue;
 import org.delia.log.Log;
 import org.delia.type.DValue;
