+++ src/test/java/org/delia/bdd/core/checker/ValueCheckerBase.java
@@ -1,4 +1,4 @@
-package org.delia.bddnew.core.checker;
+package org.delia.bdd.core.checker;
 
 import org.delia.api.DeliaSession;
