+++ src/test/java/org/delia/inputfunction/TLangProgramBuilderTests.java
@@ -4,7 +4,7 @@ import static org.junit.Assert.assertEquals;
 
 import org.delia.api.Delia;
 import org.delia.api.DeliaSession;
-import org.delia.bddnew.NewBDDBase;
+import org.delia.bdd.NewBDDBase;
 import org.delia.builder.ConnectionBuilder;
 import org.delia.builder.ConnectionInfo;
 import org.delia.builder.DeliaBuilder;
