+++ src/test/java/org/delia/type/DTypeRegistryTests.java
@@ -6,7 +6,7 @@ import static org.junit.Assert.assertEquals;
 import java.util.Set;
 
 import org.delia.api.Delia;
-import org.delia.bddnew.NewBDDBase;
+import org.delia.bdd.NewBDDBase;
 import org.delia.builder.ConnectionBuilder;
 import org.delia.builder.ConnectionInfo;
 import org.delia.builder.DeliaBuilder;
